Create README.md
This is a sample demo in how to download a zip file and save it in the public storage directory. Thereafter getting to unzip the file and save each file within the App's directory. There is a main activity in charge of listing the files and for sake of seeing file manipulation, the file can be saved, deleted, and seeing how to restore back the original zip files by repeating the process. After I read Commonsware core chapters, I saw this situation and wanted to extract that functionality and apply it in a simpler case than its EmPubLite activity.